\name{loadEChartsLibrary}
\alias{loadEChartsLibrary}

\title{
  Load the Javascript Library File of ECharts to the Shiny Application
}
\description{
  This function will help load the Javascript library file of ECharts to the current shiny project. This is mandatory before we can plot with ECharts in Shiny applications.
}
\usage{
loadEChartsLibrary()
}

\author{
  Xiaodong DENG

  (ECharts library is authored by Baidu team)
}



\examples{
# please refer to vignettes for the practical examples
loadEChartsLibrary()
}

