\name{loadWordcloudExtension}
\alias{loadWordcloudExtension}
\title{
Load the Extension for Using Word Cloud in ECharts
}
\description{
In order to use word cloud feature in this package, an extension is needed. loadWordcloudExtension() function is used to help load the extension file.

Users must run this function before they use renderWordcloud() function.
}
\usage{
loadWordcloudExtension()
}

\author{
  Xiaodong DENG

  (ECharts library is authored by Baidu team)
}


\examples{
loadWordcloudExtension() # please use this within in Shiny applications
}
